feat: Add App.Seed project with SeedDataService

+++ App.Server/src/Program.cs
@@ -1,4 +1,5 @@
-﻿
+﻿using App.Seed;
+
 var builder = WebApplication.CreateBuilder();
 
 
@@ -11,7 +12,8 @@ builder.Services.AddPooledDbContextFactory<AppDbContext>(
         options => options.UseSqlServer(appSetting.Connectionstring, sqlOptions => {
             sqlOptions.UseQuerySplittingBehavior(QuerySplittingBehavior.SplitQuery);
         }))
-    .AddScoped<AppDbContext>(p => p.GetRequiredService<IDbContextFactory<AppDbContext>>().CreateDbContext());
+    .AddScoped<AppDbContext>(p => p.GetRequiredService<IDbContextFactory<AppDbContext>>().CreateDbContext())
+    .AddScoped<SeedDataService>();
 
 builder.Services.AddGraphQLServer()
     .RegisterDbContext<AppDbContext>(DbContextKind.Pooled)
@@ -22,14 +24,25 @@ var app = builder.Build();
 
 app.MapGet("/", () => "Timesheet graphql server");
 
-app.MapPost("/ensure", async (AppDbContext context) => {
-    await context.Database.EnsureDeletedAsync();
-    await context.Database.EnsureCreatedAsync();
-    context.Activities.Add(new Activity {
-        Description = "Setup Timesheet Grraphql Server",
-        Start = new DateTimeOffset(DateTime.Now)
-    });
-    await context.SaveChangesAsync();
+
+
+
+app.MapPost("/ensure-db", async (AppDbContext context) => {
+    if (app.Environment.IsDevelopment()) {
+        await context.Database.EnsureCreatedAsync();
+    }
+});
+
+app.MapPost("/ensure-migration", async (AppDbContext context) => {
+    await context.Database.MigrateAsync();
+});
+
+app.MapPost("/seed-db", async (AppDbContext context, [Service] SeedDataService service) => {
+    if (app.Environment.IsDevelopment()) {
+        await context.Database.EnsureDeletedAsync();
+        await context.Database.EnsureCreatedAsync();
+        await service.SeedDatabase();
+    }
 });
 
 app.MapGraphQL();
